refactor: Move more parsing tests to a separate modules
The tests file take a long time to format due to its size so factoring
it out into smaller ones will reduce that (and hopefully make it clearer
where to find/put a test)
